created components (?) for FIELDS and SUBJECTS (for search)
